Added futureproof function for incorrect color
Added wait_for_seconds import and functionality that would make future color-sensing and movement easier. Having a built-in wait will make it easier and more predictable to run through a maze in the case that we begin working on that implementation. Will update color sensing for RGBi if the need arises.