Renamed file, finished color sensing
Renamed the file to be more fitting for a final project submission. Added more functionality to color sensing that will now correctly run behaviors only when it senses a color other than the color of the search area. Added an array to store the incorrect colors sensed and beeps in different tones to demonstrate memory for later usage in the maze.